Use correct VG_LIBDIR even if out is empty
If out is empty, $(realpath $(HOST_OUT_SHARED_LIBRARIES)) returns an
empty path, since it actually tries to find the path. Use abspath
instead to produce the correct result even if the path doesn't exist
yet.

Now the build.ninja is generated the same with an empty out directory as
it is during an incremental build.

Change-Id: I645aff304280d9caf180ec882073eb9cdd8edec5